+++ Tokenize/filters/ChmFilter.cc
@@ -120,7 +120,7 @@ bool ChmFilter::set_property(Properties prop_name, const string &prop_value)
 	return false;
 }
 
-bool ChmFilter::set_document_data(const char *data_ptr, unsigned int data_length)
+bool ChmFilter::set_document_data(const char *data_ptr, off_t data_length)
 {
 	return false;
 }
