+++ Tokenize/filters/ExternalFilter.cc
@@ -151,7 +151,7 @@ bool ExternalFilter::set_property(Properties prop_name, const string &prop_value
 	return true;
 }
 
-bool ExternalFilter::set_document_data(const char *data_ptr, unsigned int data_length)
+bool ExternalFilter::set_document_data(const char *data_ptr, off_t data_length)
 {
 	return false;
 }
