+++ Tokenize/filters/Filter.h
@@ -64,7 +64,7 @@ namespace Dijon
     /** Converts text to UTF-8.
      */
     typedef std::string (convert_to_utf8_func)(const char *,
-        unsigned int, const std::string &);
+        off_t, const std::string &);
 
     /// Filter interface.
     class DIJON_FILTER_EXPORT Filter
@@ -115,7 +115,7 @@ namespace Dijon
 	 * Call next_document() to position the filter onto the first document.
 	 * Returns false if this input is not supported or an error occured.
 	 */
-	virtual bool set_document_data(const char *data_ptr, unsigned int data_length) = 0;
+	virtual bool set_document_data(const char *data_ptr, off_t data_length) = 0;
 
 	/** (Re)initializes the filter with the given data.
 	 * Call next_document() to position the filter onto the first document.
