+++ Tokenize/filters/GMimeMboxFilter.cc
@@ -211,7 +211,7 @@ bool GMimeMboxFilter::set_property(Properties prop_name, const string &prop_valu
 	return false;
 }
 
-bool GMimeMboxFilter::set_document_data(const char *data_ptr, unsigned int data_length)
+bool GMimeMboxFilter::set_document_data(const char *data_ptr, off_t data_length)
 {
 	// Close/free whatever was opened/allocated on a previous call to set_document()
 	finalize(true);
