+++ Tokenize/filters/HtmlFilter.cc
@@ -583,7 +583,7 @@ bool HtmlFilter::set_property(Properties prop_name, const string &prop_value)
 	return false;
 }
 
-bool HtmlFilter::set_document_data(const char *data_ptr, unsigned int data_length)
+bool HtmlFilter::set_document_data(const char *data_ptr, off_t data_length)
 {
 	if ((data_ptr == NULL) ||
 		(data_length == 0))
