+++ Tokenize/filters/TextFilter.cc
@@ -52,7 +52,7 @@ bool TextFilter::set_property(Properties prop_name, const string &prop_value)
 	return true;
 }
 
-bool TextFilter::set_document_data(const char *data_ptr, unsigned int data_length)
+bool TextFilter::set_document_data(const char *data_ptr, off_t data_length)
 {
 	if ((data_ptr == NULL) ||
 		(data_length == 0))
