+++ Tokenize/filters/XmlFilter.cc
@@ -147,7 +147,7 @@ bool XmlFilter::set_property(Properties prop_name, const string &prop_value)
 	return true;
 }
 
-bool XmlFilter::set_document_data(const char *data_ptr, unsigned int data_length)
+bool XmlFilter::set_document_data(const char *data_ptr, off_t data_length)
 {
 	if ((data_ptr == NULL) ||
 		(data_length == 0))
